Change implementation for extracting and keeping zone data to avoid issues with instantiating too many LoadBalancerProperties beans.
